+++ src/main/java/kitchenpos/application/dto/request/ProductCommand.java
@@ -1,7 +1,7 @@
 package kitchenpos.application.dto.request;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.Product;
 
 public class ProductCommand {
