+++ src/main/java/kitchenpos/domain/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.menu;
+package kitchenpos.domain;
 
 import java.util.List;
 import kitchenpos.domain.Menu;
