+++ src/main/java/kitchenpos/domain/ProductRepository.java
@@ -1,6 +1,5 @@
 package kitchenpos.domain;
 
-import kitchenpos.domain.menu.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
