+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -3,25 +3,16 @@ package kitchenpos.application;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.application.dto.response.MenuGroupResponse;
-import kitchenpos.application.dto.response.ProductResponse;
 import kitchenpos.common.DataClearExtension;
-import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuGroupRepository;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.ProductRepository;
-import kitchenpos.domain.menu.Product;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 
 @SpringBootTest
 @DisplayName("메뉴 그룹 관련 기능에서")
@@ -33,9 +24,6 @@ class MenuGroupServiceTest {
 
     @Autowired
     private MenuGroupRepository menuGroupRepository;
-
-    @Autowired
-    private ProductRepository productRepository;
 
     @Test
     @DisplayName("메뉴 그룹을 정상적으로 생성한다.")
