+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -14,8 +14,8 @@ import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuGroupRepository;
 import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.ProductRepository;
-import kitchenpos.domain.menu.MenuRepository;
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.MenuRepository;
+import kitchenpos.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
