+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -17,8 +17,8 @@ import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.OrderTableRepository;
 import kitchenpos.domain.ProductRepository;
-import kitchenpos.domain.menu.MenuRepository;
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.MenuRepository;
+import kitchenpos.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
